pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "windforecast"
version = "0.1.0"
description = "Wind forecast analyzer and report generator for kitesurfing conditions"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Pierre Malafosse", email = "pmalafosse@example.com" }
]
dependencies = [
    "pandas>=2.0.0",
    "requests>=2.28.0",
    "pillow>=9.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
windforecast = "windforecast.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q --cov=windforecast"

[tool.black]
line-length = 100
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
